refactor(websocket): unify notification and toast systems
ARCHITECTURE CHANGES:
- NotificationFeed.vue: Remove duplicate WebSocket processing
  - Now only reads from localStorage (App.vue handles WebSocket)
  - Added architecture documentation comment
  - Removed addNotification, processMessage, debounce logic
  - Simplified to load/display notifications from centralized storage

- App.vue: Consolidate toast notification system
  - Remove redundant activeToasts/addToast/removeToast local system
  - Use unified useToast composable for all toast notifications
  - Remove ToastNotification component (ToastContainer handles display)
  - processToastNotification now uses toast.show() from useToast

- Add duplicate detection to addNotificationToStorage in App.vue
  - Prevents duplicate notifications for same events within time windows
  - stream.online/offline: 10 second window
  - channel.update: 30 second window with same title
  - recording events: 5 second window

BENEFITS:
- Single source of truth: All WebSocket messages processed in App.vue
- Notifications work even when NotificationFeed panel is closed
- Consistent toast system via useToast + ToastContainer
- No polling intervals - fully event-driven
- Reduced code duplication and complexity
